requirements.txt
discord.py
python-dotenv
pyyaml

